Use Vue Output Target (#3)
This commit adds the Vue output target to Stencil, and uses it to
create the new `vue-components` package. We're loading that package
into the Vue app to load Vue-wrapped components on the landing page.
Note the comment in `main.js` concerning a workaround for a bug in
the Vue output target.